[mod] 更改类型

+++ vercel-mongodb-count/lib/database.rs
@@ -26,8 +26,8 @@ pub async fn count(config: &DataBaseConfig) -> Result<isize, Box<dyn Error>> {
     let first_count = first.count;
     let count_documents = collection.count_documents(doc! {}, None).await?;
     let max_documents = 10;
-    let number = first.count + max_documents;
-    if count_documents > max_documents as u64 {
+    let number = first.count + max_documents as isize;
+    if count_documents > max_documents {
         let option = FindOneOptions::builder().skip(max_documents);
         let last = collection.find_one(doc! {}, option.build()).await?;
         let fid = first._id;
